+++ include/reverb.h
@@ -5,6 +5,6 @@
 extern void pinConfigReverb(void);
 extern void setUpReverb(void);
 extern void loopReverb(void);
-extern void processReverbAudio(int inputSample); 
+extern void processReverbAudio(int16_t inputSample);
 
 #endif
